pyproject.toml
[project]
name = "dollardollar"
version = "0.1.0"
requires-python = ">3.11"
readme = "README.md"
dependencies = [
    "alembic>=1.7.5",
    "cryptography==41.0.5",
    "flask==3.1.0",
    "flask-apscheduler>=1.13.1",
    "flask-login==0.6.3",
    "flask-mail==0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn==23.0.0",
    "oauthlib==3.2.2",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.8.0",
    "pyopenssl==23.2.0",
    "python-dotenv==1.0.1",
    "pytz>=2021.1",
    "requests==2.32.3",
    "requests-oauthlib==1.3.1",
    "ruff>=0.11.6",
    "schedule>=1.2.2",
    "scrypt>=0.8.27",
    "tzdata>=2021.1",
    "werkzeug==3.1.3",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "BLE",    # flake8-blind-except
    "C",      # complexity
    "COM",    # flake8 commas
    "C90",    # mccabe
    "D",      # docstrings
    "DTZ",    # flake8-datetimez
    "E",      # pycodestyle
    "F",      # pyflakes/autoflake
    "FLY",    # flynt
    "FURB",   # refurb
    "G",      # flake8-logging-format
    "I",      # isort
    "INP",    # flake8-no-pep420
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # import conventions;
    "LOG",    # flake8-logging
    "N",      # pep8-naming
    "PERF",   # Perflint
    "PGH",    # pygrep-hooks
    "PIE",    # flake8-pie
    "PL",     # pylint
    "PT",     # flake8-pytest-style
    "PYI",    # flake8-pyi
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # Ruff specific
    "S",      # flake8-bandit
    "SIM",    # flake8-simplify
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    "TID251", # Banned imports
    "TRY",    # tryceratops
    "UP",     # pyupgrade
    "W",      # pycodestyle
]

ignore = [
    "D100",   # Missing docstring in public module,
    "D103",   # Missing docstring in public function
    "D203",   # 1 blank line before class docstring
    "D211",   # No blank lines allowed before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
    "INP001", # File is part of an implicit namespace package. Add an __init__.py

    # May conflict with the formatter, https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "PLR2004",
    "Q",
    "COM812",
    "COM819",

    # Disabled because ruff does not understand type of __all__ generated by a function
    "PLE0605",

    "PLR0912", # too many functions and too complex to fix this for now
    "C901",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["ipython>=9.4.0"]
